fix: resolve -Wdouble-promotion and -Wunused-variable warnings
Fix float-to-double implicit conversion warnings in BNO08x.c and BNO08x_SPI.c LOG_INF calls by casting to double. Remove unused current_rx_ticks variable in esb.c.

+++ src/sensor/imu/BNO08x.c
@@ -412,7 +412,7 @@ int bno08x_update_odr(float accel_time, float gyro_time,
         bno.gyro_time = bno.actual_time;
         *accel_actual_time = bno.actual_time;
         *gyro_actual_time = bno.actual_time;
-        LOG_INF("ODR updated to %.1f Hz", 1.0f / bno.actual_time);
+        LOG_INF("ODR updated to %.1f Hz", (double)(1.0f / bno.actual_time));
     }
     k_mutex_unlock(&bno_mutex);
     return ret;
